+++ src/Services/StateActionsService.php
@@ -2,59 +2,10 @@
 
 namespace Abather\SpatieLaravelModelStatesActions\Services;
 
-use Abather\SpatieLaravelModelStatesActions\Traits\Makeable;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Foundation\Auth\User;
+use Filament\Tables\Filters\SelectFilter;
 
-class StateActionsService
+class StateActionsService extends _BaseService
 {
-    use Makeable;
-
-    /** @var Model::class */
-    private string $model;
-
-    private string $field;
-
-    private ?User $user;
-
-    private array $excluded_states = [];
-
-    private array $include_states = [];
-
-    public function __construct(string $model, string $field = 'state', ?User $user = null)
-    {
-        $this->model = $model;
-        $this->field = $field;
-        $this->user = $user ?? auth()->user();
-    }
-
-    public function excludeStates(array|string $excluded_states): self
-    {
-        if (is_string($excluded_states)) {
-            $excluded_states = [$excluded_states];
-        }
-
-        $this->excluded_states = array_merge($this->excluded_states, $excluded_states);
-
-        return $this;
-    }
-
-    public function includeStates(array|string $include_states): self
-    {
-        if (is_string($include_states)) {
-            $include_states = [$include_states];
-        }
-
-        $this->include_states = array_merge($this->include_states, $include_states);
-
-        return $this;
-    }
-
-    private function getStates(): array
-    {
-        return array_diff($this->model::getStatesFor($this->field)->toArray(), $this->excluded_states);
-    }
-
     public function tableActions(): array
     {
         $actions = [];